Polymarket-btc-bot/requirements.txt
# requirements.txt
websocket-client
requests
python-dotenv
py-clob-client
web3